requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14,<4.0
django-filter>=23.0
